requirements.txt
tweetnlp
streamlit
plotly-express
git+https://github.com/JustAnotherArchivist/snscrape.git